rm(all): Remove all debug print + create file for test
